+++ src/main/java/com/jasik/momsnaggingapi/domain/nagging/controller/NaggingController.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.nagging.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class NaggingController {
+}
