+++ src/main/java/com/jasik/momsnaggingapi/domain/nagging/repository/NaggingRepository.java
@@ -1,0 +1,7 @@
+package com.jasik.momsnaggingapi.domain.nagging.repository;
+
+import com.jasik.momsnaggingapi.domain.nagging.Nagging;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NaggingRepository extends JpaRepository<Nagging, Long> {
+}
